+++ apps/ledger-live-mobile/src/screens/Swap/LiveApp/customHandlers/index.ts
@@ -152,7 +152,7 @@ export function useSwapCustomHandlers(
     "custom.getFee": getFee(accounts, navigation),
     "custom.getTransactionByHash": getTransactionByHash(accounts),
     "custom.saveSwapToHistory": saveSwapToHistory(accounts, dispatch),
-    "custom.swapRedirectToHistory": () => null,
+    "custom.swapRedirectToHistory": () => navigation.navigate(ScreenName.SwapHistory),
     "custom.isReady": async () => {
       if (Config.DETOX) {
         sendSwapLiveAppReady();