+++ e2e/mobile/page/discover/discover.page.ts
@@ -16,7 +16,7 @@ export default class DiscoverPage {
     { name: "Zerion", url: " https://zerion.io/" },
     { name: "Rainbow", url: " https://rainbow.me" },
     { name: "POAP", url: " https://app.poap.xyz/" },
-    { name: "Yearn", url: " https://beta.yearn.finance" },
+    { name: "Yearn", url: " https://beta.yearn.fi" },
     { name: "ChangeNOW", url: " https://changenow.io/" },
     { name: "Transak", url: " https://transak.com" },
   ];