+++ e2e/mobile/specs/swap/otherTestCases/swapLandingPage.spec.ts
@@ -4,11 +4,12 @@ const swapTestConfig = {
   fromAccount: Account.ETH_1,
   toAccount: TokenAccount.ETH_USDC_1,
   tmsLinks: ["B2CQA-2918"],
+  tags: ["@NanoSP", "@LNS", "@NanoX"],
 };
 
 runSwapLandingPageTest(
   swapTestConfig.fromAccount,
   swapTestConfig.toAccount,
   swapTestConfig.tmsLinks,
-  ["@NanoSP", "@LNS", "@NanoX"],
+  swapTestConfig.tags,
 );