+++ e2e/mobile/specs/swap/otherTestCases/tooLowAmountForQuoteSwaps/EthToBtc10000NotEnoughFee.spec.ts
@@ -1,0 +1,19 @@
+import { runTooLowAmountForQuoteSwapsTest } from "../swap.other";
+
+const transactionE2E = {
+  swap: new Swap(Account.ETH_1, Account.BTC_NATIVE_SEGWIT_1, "10000"),
+  tmsLinks: ["B2CQA-3243"],
+  errorMessage: "Not enough balance, including network fee",
+  ctaBanner: true,
+  quotesVisible: false,
+  tags: ["@NanoSP", "@LNS", "@NanoX"],
+};
+
+runTooLowAmountForQuoteSwapsTest(
+  transactionE2E.swap,
+  transactionE2E.tmsLinks,
+  transactionE2E.errorMessage,
+  transactionE2E.ctaBanner,
+  transactionE2E.quotesVisible,
+  transactionE2E.tags,
+);