+++ libs/device-react/CHANGELOG.md
@@ -1,11 +1,18 @@
 # @ledgerhq/device-react
 
-## 0.2.34-nightly.0
+## 0.2.34
 
 ### Patch Changes
 
 - Updated dependencies []:
-  - @ledgerhq/device-core@0.4.28-nightly.0
+  - @ledgerhq/device-core@0.4.28
+
+## 0.2.34-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/device-core@0.4.28-next.0
 
 ## 0.2.33