+++ libs/ledgerjs/packages/hw-app-near/CHANGELOG.md
@@ -1,11 +1,18 @@
 # @ledgerhq/hw-app-near
 
-## 6.31.3-nightly.0
+## 6.31.3
 
 ### Patch Changes
 
 - Updated dependencies []:
-  - @ledgerhq/hw-transport@6.31.7-nightly.0
+  - @ledgerhq/hw-transport@6.31.7
+
+## 6.31.3-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-transport@6.31.7-next.0
 
 ## 6.31.2