+++ libs/ledgerjs/packages/hw-bolos/CHANGELOG.md
@@ -1,11 +1,18 @@
 # @ledgerhq/hw-bolos
 
-## 6.32.3-nightly.0
+## 6.32.3
 
 ### Patch Changes
 
 - Updated dependencies []:
-  - @ledgerhq/hw-transport@6.31.7-nightly.0
+  - @ledgerhq/hw-transport@6.31.7
+
+## 6.32.3-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-transport@6.31.7-next.0
 
 ## 6.32.2