+++ libs/live-dmk-shared/CHANGELOG.md
@@ -1,11 +1,18 @@
 # @ledgerhq/live-dmk
 
-## 0.9.1-nightly.0
+## 0.9.1
 
 ### Patch Changes
 
 - Updated dependencies []:
-  - @ledgerhq/hw-transport@6.31.7-nightly.0
+  - @ledgerhq/hw-transport@6.31.7
+
+## 0.9.1-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-transport@6.31.7-next.0
 
 ## 0.9.0